cc warning

+++ avbsf/src/h264bsf.c
@@ -76,7 +76,7 @@ static int h264bsf_input(void* param, int64_t pts, int64_t dts, const uint8_t* n
 
 	if (bsf->vcl && (dts != bsf->dts || 0 == bytes || h264_is_new_access_unit(nalu, bytes)))
 	{
-		r = bsf->onpacket(bsf->param, bsf->pts, bsf->dts, bsf->ptr.ptr, bsf->ptr.len, 1==bsf->vcl ? 0x01 : 0);
+		r = bsf->onpacket(bsf->param, bsf->pts, bsf->dts, bsf->ptr.ptr, (int)bsf->ptr.len, 1==bsf->vcl ? 0x01 : 0);
 		bsf->ptr.len = 0;
 		bsf->sps_pps_flag = 0;
 		bsf->vcl = 0;
